+++ ui/App.vue
@@ -6,7 +6,6 @@
         </main>
         <main-footer></main-footer>
         <audio id="sharp" src="/sounds/sharp.ogg" crossOrigin="anonymous"></audio>
-        <update-installed-toast></update-installed-toast>
         <app-update-available></app-update-available>
         <communication-error></communication-error>
         <dataset-import-failure></dataset-import-failure>