+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 - 0.2.6-beta
-    - Added export chart feature to Web UI
-    - Add web app update notification
+    - Added export chart feature to web UI
+    - Added web UI automatic updates
 
 - 0.2.5-beta
     - Added Backoff and Retry client middleware